Improve ledger::block_destination and rename to ledger::destination
The new implementation queries block::type () and uses block sideband to determine if a block is a send and therefore has a destination. It also uses std::optional rather than a sentinel value to indicate if a block has no destination.
Previously the implementation was using two dynamic casts and querying the ledger::is_send to determine if a block had a destination. It was also using a sentinel value that wasn't being checked properly in all instances.